App/CreateRoomScreen - + moved 'creator' components directory from "models" to "components" dir.
